move test to cause-rendering
